requirements.txt
# ViM 无人机射频开集识别 - 依赖包列表
# 安装命令：pip install -r requirements.txt

# ---- 深度学习核心 ----
torch>=2.0.0
torchvision>=0.15.0

# ---- 科学计算 ----
numpy>=1.24.0
scipy>=1.10.0

# ---- 机器学习工具 ----
scikit-learn>=1.2.0

# ---- 配置文件解析 ----
pyyaml>=6.0

# ---- 训练可视化 ----
tensorboard>=2.12.0

# ---- 进度条 ----
tqdm>=4.65.0

# ---- 数据处理 ----
Pillow>=9.5.0
